08-capstone-project/requirements.txt
airflow-dag==0.1.0
boto3==1.26.27
botocore==1.29.27
click==8.1.3
Jinja2==3.1.2
jmespath==1.0.1
MarkupSafe==2.1.1
python-dateutil==2.8.2
PyYAML==6.0
s3transfer==0.6.0
six==1.16.0
urllib3==1.26.13
